--- src/test/java/school/redrover/CreateFreestyleProjectTest_2.java
@@ -1,36 +1,0 @@
-package school.redrover;
-
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-import school.redrover.runner.BaseTest;
-
-public class CreateFreestyleProjectTest_2 extends BaseTest {
-
-    @Test
-    public void testAddingNewItem() {
-        final String nameOfJob = "Katya's Project";
-        final String expectedH1 = "Project Katya's Project";
-
-        WebElement newItemMenu = getWait2().until(ExpectedConditions.elementToBeClickable(By.xpath("//a[@href = '/view/all/newJob']")));
-        newItemMenu.click();
-
-        WebElement nameInputField = getDriver().findElement(By.id("name"));
-        nameInputField.sendKeys(nameOfJob);
-
-        WebElement freestyleProjectButton = getDriver().findElement(By.xpath("//li//span[contains(text(), 'Freestyle')]"));
-        freestyleProjectButton.click();
-
-        WebElement okButton = getDriver().findElement(By.id("ok-button"));
-        okButton.click();
-
-        WebElement saveButton = getDriver().findElement(By.name("Submit"));
-        saveButton.click();
-
-        WebElement myViewsMenuItem = getWait2().until(ExpectedConditions.presenceOfElementLocated(By.xpath("//h1")));
-
-        Assert.assertEquals(myViewsMenuItem.getText(), expectedH1);
-    }
-}
